requirements.txt
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
blinker==1.4
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.1
MarkupSafe==1.1.1
passlib==1.7.4
pycparser==2.20
python-dotenv==0.15.0
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
speaklater==1.3
SQLAlchemy==1.3.19
urllib3==1.25.11
Werkzeug==1.0.1
WTForms==2.3.3
